refactor static files handler
